mir: centralize expression class derivation
